Move json to utils
