+++ app/utils/json.go
@@ -1,37 +1,14 @@
-package handlers
+package utils
 
 import (
 	"encoding/json"
 	"errors"
 	"io"
 	"net/http"
-	"net/http/httptest"
-	"testing"
-
-	"github.com/go-chi/chi/v5"
 )
 
-// executeRequest, creates a new ResponseRecorder
-// then executes the request by calling ServeHTTP in the router
-// after which the handler writes the response to the response recorder
-// which we can then inspect.
-func executeRequest(req *http.Request, router *chi.Mux) *httptest.ResponseRecorder {
-	rr := httptest.NewRecorder()
-	router.ServeHTTP(rr, req)
-
-	return rr
-}
-
-// checkResponseCode is a simple utility to check the response code
-// of the response
-func checkResponseCode(t *testing.T, expected, actual int) {
-	if expected != actual {
-		t.Errorf("Expected response code %d. Got %d\n", expected, actual)
-	}
-}
-
-
-func writeJSON(w http.ResponseWriter, status int, data interface{}, wrap ...string) error {
+// WriteJSON writes the data to the response writer as json
+func WriteJSON(w http.ResponseWriter, status int, data interface{}, wrap ...string) error {
 	// out will hold the final version of the json to send to the client
 	var out []byte
 
@@ -66,7 +43,8 @@ func writeJSON(w http.ResponseWriter, status int, data interface{}, wrap ...stri
 	return nil
 }
 
-func errorJSON(w http.ResponseWriter, err error, status ...int) {
+// ErrorJSON writes an error to the response writer as json
+func ErrorJSON(w http.ResponseWriter, err error, status ...int) {
 	statusCode := http.StatusBadRequest
 	if len(status) > 0 {
 		statusCode = status[0]
@@ -80,10 +58,11 @@ func errorJSON(w http.ResponseWriter, err error, status ...int) {
 		Message: err.Error(),
 	}
 
-	_ = writeJSON(w, statusCode, theError, "error")
+	_ = WriteJSON(w, statusCode, theError, "error")
 }
 
-func readJSON(w http.ResponseWriter, r *http.Request, data interface{}) error {
+// ReadJSON reads the request body and decodes the json into the data interface
+func ReadJSON(w http.ResponseWriter, r *http.Request, data interface{}) error {
 	maxBytes := 1024 * 1024 // one megabyte
 	r.Body = http.MaxBytesReader(w, r.Body, int64(maxBytes))
 	dec := json.NewDecoder(r.Body)
